small fix on http interface

+++ src/main/java/com/catJam/Order/bookClient/BookClient.java
@@ -15,7 +15,7 @@ public interface BookClient {
     List<BookEntity> findAll();
 
     @GetExchange("/books/{id}")
-    BookEntity findById(Integer id);
+    BookEntity findById(Long id);
 
     @PostExchange("/books")
     BookEntity create(BookEntity bookEntity);
